requirements.txt
# Django stuff
Django==1.6.2

# Timezone support
pytz==2014.2

# PIL support
Pillow==2.4.0

# jinja2 templates
Jinja2==2.7.2
git+ssh://git@github.com/concentricsky/jingo.git@v0.4

# mysql database
mysql-python==1.2.5

# migrations
South==0.8.4

# CSky models
git+ssh://git@github.com/concentricsky/django-cachemodel.git@v2.0.0
git+ssh://git@github.com/concentricsky/django-basic-models.git@v2.0.1

# CSky Sphinx search lib
git+https://github.com/concentricsky/django-sphinx.git@v2.2.7

# CSky Client Admin and CMS
git+ssh://git@github.com/concentricsky/django-skycms.git@v0.7.0
git+ssh://git@github.com/concentricsky/django-sky-ckeditor.git@v0.9.8
git+ssh://git@github.com/concentricsky/django-client-admin.git@v1.0.14
git+ssh://git@github.com/concentricsky/django-skythumbnails.git@v0.2.5

git+ssh://git@github.com/concentricsky/django-sky-redirects.git@v0.8.1

# CSky visitor
git+ssh://git@github.com/concentricsky/django-emailtemplates.git@v0.10.4
git+ssh://git@github.com/concentricsky/django-sky-visitor.git@v1.1.1

# Revision and deleted object recovery
django-reversion==1.8.0

# Latest Rest Framework as of 2014-04-12
djangorestframework==2.3.13
Markdown==2.4